+++ src/us/kbase/workspace/test/kbase/JSONRPCLayerTest.java
@@ -101,7 +101,7 @@ import com.google.common.collect.ImmutableMap;
  */
 public class JSONRPCLayerTest extends JSONRPCLayerTester {
 
-	private static final String VER = "0.14.3";
+	private static final String VER = "0.15.0";
 
 	@Test
 	public void ver() throws Exception {